Add ayacript plugin.
